// apps/web/src/lib/api.gen.ts
export interface SignUpWith2FAResponse {
  qr_image: string;
  secret: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface ChangePassword {
  oldPassword: string;
  newPassword: string;
  /** @pattern ^\d{6}$ */
  twoFACode?: string;
}

export interface LoginSuccessWithJWT {
  accessToken: string;
  refreshToken: string;
}

export type LoginRequires2FAType = typeof LoginRequires2FAType[keyof typeof LoginRequires2FAType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LoginRequires2FAType = {
  '2fa': '2fa',
} as const;

export interface LoginRequires2FA {
  type: LoginRequires2FAType;
  /** Temporary token to use for 2FA verification */
  temp_token: string;
}

export interface Refresh {
  refreshToken: string;
}

export interface RevokeCurrentSession {
  refreshToken: string;
}

export interface GetAllSessionResponse {
  issuedAt: string;
  expiresAt: string;
  ipAddress: string;
  userAgent: string;
  createdBy: string;
}

export interface TwoFAResponse {
  qr_image: string;
  secret: string;
}
// ...
export interface RoleResponse {
  id: number;
  /** @pattern [a-zA-Z_-]+ */
  name: string;
  description: string;
  createdAt: string;
  createdBy: string;
  updatedAt: string;
  updatedBy: string;
}

export interface PermissionResponse {
  id: number;
  /** @pattern [a-zA-Z_-]+ */
  name: string;
  description: string;
  createdAt: string;
  createdBy: string;
  updatedAt: string;
  updatedBy: string;
}

export interface AssignPermission {
  ids: number[];
}

export type RolesAndPermissionResponseRoles = {
  id: number;
  /** @pattern [a-zA-Z_-]+ */
  name: string;
  description: string;
  createdAt: string;
  createdBy: string;
  updatedAt: string;
  updatedBy: string;
  permissions: PermissionResponse[];
};

export interface RolesAndPermissionResponse {
  roles: RolesAndPermissionResponseRoles;
}

export interface UserWithRoles {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  roles: string[];
  permissions: string[];
}

export interface AssignRoleToUser {
  roles: number[];
}

export interface UserDetails {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  locked: boolean;
  disabled: boolean;
  twoFAEnabled: boolean;
  createdAt: string;
  updatedAt: string;
  roles: string[];
  permissions: string[];
}
// ...
export type GetListOfUsersParams = {
/**
 * Enter the first name of the user you want to search
 */
firstName?: string;
/**
 * Enter the first name of the user you want to search
 */
lastName?: string;
/**
 * Enter the first name of the user you want to search
 */
email?: string;
/**
 * Enter the first name of the user you want to search
 */
size?: number;
/**
 * Enter the first name of the user you want to search
 */
page?: number;
};